Adjust Object's update to recv int64

+++ world/Object.go
@@ -18,7 +18,7 @@ type Object struct {
 }
 
 // update updates the given object.
-func (o *Object) update(d int) {
+func (o *Object) update(d int64) {
 }
 
 // GetOwner returns the owning object.
